api/requirements.txt
flask==3.1.0
Werkzeug==3.1.3
pillow==11.1.0
flask-cors==5.0.0
gunicorn==23.0.0
PyQt6==6.4.2